# Conflicts:
#	BackEnd/QArte/QArte.API/Controllers/JwtSecutirtyTokenHandler.cs
#	BackEnd/QArte/QArte.API/Program.cs
#	BackEnd/QArte/QArte.API/QArte.API.csproj
#	BackEnd/QArte/QArte.API/appsettings.json
#	BackEnd/QArte/QArte.Persistance/Migrations/20240129111821_MyMigration.Designer.cs
#	BackEnd/QArte/QArte.Persistance/Migrations/20240129111821_MyMigration.cs
#	BackEnd/QArte/QArte.Persistance/Migrations/20240205123307_MyMigration.Designer.cs
#	BackEnd/QArte/QArte.Persistance/Migrations/20240205123307_MyMigration.cs
#	BackEnd/QArte/QArte.Persistance/Migrations/20240207081356_MyMigration.Designer.cs
#	BackEnd/QArte/QArte.Persistance/Migrations/20240207081356_MyMigration.cs
#	BackEnd/QArte/QArte.Persistance/QArteDBContext.cs
#	BackEnd/QArte/QArte.Services/Services/GoogleAccuntService.cs
#	FrontEnd/QArte/src/App.tsx
#	FrontEnd/package-lock.json

+++ BackEnd/QArte/QArte.Persistance/PersistanceModels/Fee.cs
@@ -4,9 +4,9 @@ namespace QArte.Persistance.PersistanceModels
 	public class Fee
 	{
 		public int ID { get; set; }
-		public decimal Amount { get; set; }
+		public long Amount { get; set; }
 		public string Currency { get; set; }
-		public decimal ExchangeRate { get; set; }
+		public double ExchangeRate { get; set; }
 
 	}
 }
